Modifications to module hdmi_rpi
This commit changes the module's name, cleans up the Makefile,
adds casio IR codes received from IR receivers.

This commit also adds scripts/at.sh, which is to be placed in
the /home/pi in the Raspberry pi when testing this, because the
rule assumes that at.sh is present in the /home/pi directory.